Rename file_opener.py to file_utils.py
