+++ swiftsimio/file_utils.py
