+++ swiftsimio/metadata/objects.py
@@ -14,7 +14,7 @@ import h5py
 from swiftsimio.conversions import swift_cosmology_to_astropy
 from swiftsimio import metadata
 from swiftsimio.objects import cosmo_array, cosmo_quantity, cosmo_factor
-from swiftsimio.opener import FileOpener
+from swiftsimio.file_utils import FileOpener
 from abc import ABC, abstractmethod
 
 import re
@@ -819,7 +819,7 @@ class SWIFTUnits(object):
     """
 
     def __init__(
-        self, filename: Path, opener: FileOpener, handle: Optional[h5py.File] = None
+        self, filename: Path, file_opener: FileOpener, handle: Optional[h5py.File] = None
     ):
         """
         SWIFTUnits constructor
@@ -838,7 +838,7 @@ class SWIFTUnits(object):
 
         """
         self.filename = filename
-        self.opener = opener
+        self.file_opener = file_opener
         self._handle = handle
 
         self.get_unit_dictionary()
@@ -852,7 +852,7 @@ class SWIFTUnits(object):
         with other objects for efficiency reasons.
         """
         if not self._handle:  # if self._handle is None, or if file closed (h5py #1363)
-            self._handle = self.opener.open(self.filename, "r")
+            self._handle = self.file_opener.open(self.filename, "r")
 
         return self._handle
