+++ swiftsimio/reader.py
@@ -28,7 +28,7 @@ from typing import Union, List
 
 def generate_getter(
     filename,
-    opener,
+    file_opener,
     name: str,
     field: str,
     unit: unyt.unyt_quantity,
@@ -133,7 +133,7 @@ def generate_getter(
         if current_value is not None:
             return current_value
         else:
-            with opener.open(filename, "r") as handle:
+            with file_opener.open(filename, "r") as handle:
                 try:
                     if mask is not None:
                         output_type = handle[field].dtype
@@ -381,7 +381,7 @@ class __SWIFTNamedColumnDataset(object):
         return self.named_columns == other.named_columns and self.name == other.name
 
 
-def generate_datasets(group_metadata: SWIFTGroupMetadata, opener, mask):
+def generate_datasets(group_metadata: SWIFTGroupMetadata, file_opener, mask):
     """
     Generates a SWIFTGroupDatasets _class_ that corresponds to the
     particle type given.
@@ -463,7 +463,7 @@ def generate_datasets(group_metadata: SWIFTGroupMetadata, opener, mask):
             field_property = property(
                 generate_getter(
                     filename,
-                    opener,
+                    file_opener,
                     field_name,
                     field_path,
                     unit=field_unit,
@@ -492,7 +492,7 @@ def generate_datasets(group_metadata: SWIFTGroupMetadata, opener, mask):
                 this_named_column_dataset_dict[column] = property(
                     generate_getter(
                         filename,
-                        opener,
+                        file_opener,
                         column,
                         field_path,
                         unit=field_unit,
@@ -560,7 +560,7 @@ class SWIFTDataset(object):
         are specified in metadata.particle_types.
     """
 
-    def __init__(self, filename, opener, mask=None):
+    def __init__(self, filename, file_opener, mask=None):
         """
         Constructor for SWIFTDataset class
 
@@ -572,7 +572,7 @@ class SWIFTDataset(object):
             mask object containing dataset to selected particles
         """
         self.filename = filename
-        self.opener = opener
+        self.file_opener = file_opener
         self.mask = mask
 
         if mask is not None:
@@ -603,7 +603,7 @@ class SWIFTDataset(object):
         this function again if you mess things up.
         """
 
-        self.units = SWIFTUnits(self.filename, self.opener)
+        self.units = SWIFTUnits(self.filename, self.file_opener)
 
         return
 
@@ -636,7 +636,7 @@ class SWIFTDataset(object):
                 group_name,
                 generate_datasets(
                     getattr(self.metadata, f"{group_name}_properties"),
-                    self.opener,
+                    self.file_opener,
                     self.mask,
                 ),
             )
